+++ src/main/java/com/besson/arknights/block/custom/workstation/MaterialCrates.java
@@ -1,0 +1,34 @@
+package com.besson.arknights.block.custom.workstation;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModAbstractChestBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.function.Supplier;
+
+public class MaterialCrates extends ModAbstractChestBlock<MaterialCratesBE> {
+
+    private static final VoxelShape SHAPE = Block.createCuboidShape(0, 0, 0, 16, 24, 16);
+
+    public MaterialCrates(Settings settings, Supplier<BlockEntityType<? extends MaterialCratesBE>> blockEntityTypeSupplier) {
+        super(settings, blockEntityTypeSupplier);
+    }
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+
+    @Override
+    public @Nullable BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
+        return new MaterialCratesBE(pos, state);
+    }
+}